Reorganize client-area files  - add subfolders for each section  - rename authentication to clientAreaLogin
