+++ client/src/views/client-area/ClientAreaPage.vue
@@ -85,13 +85,13 @@ import ClientAreaSidebar from '@/views/client-area/ClientAreaSidebar.vue';
 import { BmsAccessInstance, BmsOrganization, DataType } from '@/services/bms';
 import { onMounted, onUnmounted, ref, watch } from 'vue';
 import { ClientAreaPages } from '@/views/client-area/types';
-import BillingDetailsPage from '@/views/client-area/BillingDetailsPage.vue';
-import ContractsPage from '@/views/client-area/ContractsPage.vue';
-import DashboardPage from '@/views/client-area/DashboardPage.vue';
-import InvoicesPage from '@/views/client-area/InvoicesPage.vue';
-import PaymentMethodsPage from '@/views/client-area/PaymentMethodsPage.vue';
-import PersonalDataPage from '@/views/client-area/PersonalDataPage.vue';
-import StatisticsPage from '@/views/client-area/StatisticsPage.vue';
+import BillingDetailsPage from '@/views/client-area/billing-details/BillingDetailsPage.vue';
+import ContractsPage from '@/views/client-area/contracts/ContractsPage.vue';
+import DashboardPage from '@/views/client-area/dashboard/DashboardPage.vue';
+import InvoicesPage from '@/views/client-area/invoices/InvoicesPage.vue';
+import PaymentMethodsPage from '@/views/client-area/payment-methods/PaymentMethodsPage.vue';
+import PersonalDataPage from '@/views/client-area/personal-data/PersonalDataPage.vue';
+import StatisticsPage from '@/views/client-area/statistics/StatisticsPage.vue';
 import useSidebarMenu from '@/services/sidebarMenu';
 import { Translatable } from 'megashark-lib';
 import { ClientAreaQuery, getCurrentRouteQuery, navigateTo, Routes } from '@/router';
