+++ client/src/views/client-area/contracts/ContractsPage.vue
