+++ client/src/views/client-area/dashboard/DashboardPage.vue
