+++ client/src/views/client-area/statistics/StatisticsPage.vue
